requirements.txt
torch
torchvision
librosa
matplotlib
numpy
Pillow